+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-dev:
+1.28.1:
   - generate error message if "validator credentials set" process fails to generate any credentials
   - allow import of accounts with null name field in their keystore
   - show text of execution payload extra data if available